% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeneticTree.R
\name{createTree}
\alias{createTree}
\title{Create a neighbor join tree object}
\usage{
createTree(alignment)
}
\arguments{
\item{alignment}{A MsaAAMultipleAlignment object for creating phylogenetic
tree}
}
\value{
Returns a phylo object which is a phylogenetic tree
}
\description{
A function that creates a neighbor join tree from multiple sequence alignment
}
\examples{
# Example 1
# Create a basic msa and then create a tree from it
package(msa)
package(Biostrings)
set1 <- Biostrings::AAStringSet("ATCGATCG")
set2 <- Biostrings::AAStringSet("ATTTTTTT")
set3 <- Biostrings::AAStringSet("ATCGATTT")
set <- union(set1, set2)
set <- union(set, set3)
align <- msa::msa(set)
createTree(align)

}
\references{
Charif D, Lobry J. 2007. “SeqinR 1.0-2: a contributed package to the R
project for statistical computing devoted to biological sequences retrieval
and analysis.” In Bastolla U, Porto M, Roman H, Vendruscolo M (eds.),
Structural approaches to sequence evolution: Molecules, networks,
populations, series Biological and Medical Physics, Biomedical Engineering,
207-232. Springer Verlag, New York.

Paradis E. & Schliep K. 2019. ape 5.0: an environment for modern
phylogenetics and evolutionaryanalyses in R. Bioinformatics 35: 526-528.

U. Bodenhofer, E. Bonatesta, C. Horejs-Kainrath, and S. Hochreiter (2015) msa:
an R package for multiple sequence alignment. Bioinformatics 31(24):3997-
9999. DOI: 10.1093/bioinformatics/btv176.
}
