% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{getSequenceByRegion}
\alias{getSequenceByRegion}
\title{Read genome sequence data from user by inputing region}
\usage{
getSequenceByRegion(region)
}
\arguments{
\item{region}{A string indicating the interested region in which SARS-COV2
are located that you want to study}
}
\value{
Returns a AAStringSet with sequences.
}
\description{
A function that takes region as input. It will look up the accessionID of
SARS-COV2 of this region. Then it will retrieve the sequence from NCBI
database online and return as a AAStringSet object.
}
\examples{
# Example 1
# Load MT066156.1 (accessionID of SARS-COV2 in Italy) by
# using "Italy" as input
region <- "Italy"
italySequence <- getSequenceByRegion("Italy")

}
\references{
H. Pagès, P. Aboyoun, R. Gentleman and S. DebRoy (2020). Biostrings:
Efficient manipulation of biological strings.
R package version 2.58.0. https://bioconductor.org/packages/Biostrings

Paradis E. & Schliep K. 2019. ape 5.0: an environment for modern
phylogenetics and evolutionary
analyses in R. Bioinformatics 35: 526-528.
}
