% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceComparison.R
\name{plotAlignment}
\alias{plotAlignment}
\title{Plot multiple sequence alignments}
\usage{
plotAlignment(alignment, refid, startIdx, endIdx)
}
\arguments{
\item{alignment}{A MsaAAMultipleAlignment object which is a result of multiple sequence alignment}

\item{refid}{A string indicating the name of reference sequence. This sequence must be in the alignment}

\item{startIdx}{A integer indicating the start position of alignment that user wants to plot}

\item{endIdx}{A integer indicating the end position of alignment that user wants to plot}
}
\value{
Returns a plot
}
\description{
A function that plot a heatmap of multiple sequence alignments, showing all
nucleotide that is different from the reference sequence
}
\examples{
# Example 1
# Create a basic msa and then plot a tree from it
library(Biostrings)
set1 <- Biostrings::DNAStringSet("ATCGATCG")
set2 <- Biostrings::DNAStringSet("ATTTTTTT")
set3 <- Biostrings::DNAStringSet("ATCGATTT")
setTotal <- union(set1, set2)
setTotal <- union(setTotal, set3)
names(setTotal) <- c("a", "b", "c")
align <- multipleSeqAlign(setTotal)
plotAlignment(align, refid = "a", startIdx = 1, endIdx = 8)

}
\references{
Charif D, Lobry J. 2007. “SeqinR 1.0-2: a contributed package to the R
project for statistical computing devoted to biological sequences retrieval
and analysis.” In Bastolla U, Porto M, Roman H, Vendruscolo M (eds.),
Structural approaches to sequence evolution: Molecules, networks,
populations, series Biological and Medical Physics, Biomedical Engineering,
207-232. Springer Verlag, New York.

Raivo Kolde (2019). pheatmap: Pretty Heatmaps. R package version 1.0.12.
https://CRAN.R-project.org/package=pheatmap
}
