% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readNameToRegions}
\alias{readNameToRegions}
\title{Read a name to region dataframe. A helper function for readGenome}
\usage{
readNameToRegions(nameToRegionsFile)
}
\arguments{
\item{nameToRegionsFile}{A string indicating the path of name to region .csv file}
}
\value{
Returns a DNAStringSet with sequences from fastaFile are add.
}
\description{
A function that read name to region dataframe from user by inputting
directories.
}
\references{
H. Pagès, P. Aboyoun, R. Gentleman and S. DebRoy (2020). Biostrings:
Efficient manipulation of biological strings.
R package version 2.58.0. https://bioconductor.org/packages/Biostrings
}
\keyword{internal}
