% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCov2Comparator.R
\name{runCov2Comparator}
\alias{runCov2Comparator}
\title{Launch Shiny App For Package Cov2Comparator}
\usage{
runCov2Comparator()
}
\value{
No return value; a shiny page is opened.
}
\description{
A function that launches the shiny app for this package.
This shiny app allows user to input sequence .fasta files,
or select genome in the menu. Then a multiple sequence
alignment and a phylogenetic tree will be calculated.
Plots for multiple sequence alignment and phylogenetic tree
will be presented.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
cov2Comparator::runCov2Comparator()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
Steipe B., ABC project (.utility 4.07) A Bioinformatics Course: Applied Bioinformatics
\href{http://steipe.biochemistry.utoronto.ca/abc/index.php/Bioinformatics_Main_Page}{Link}.
}
