% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceComparison.R
\name{saveAlignmentToFasta}
\alias{saveAlignmentToFasta}
\title{Save alignment to a fasta file}
\usage{
saveAlignmentToFasta(alignment, outputName)
}
\arguments{
\item{alignment}{A MsaAAMultipleAlignment obtained from multiple sequence
alignment}

\item{outputName}{A string indicating the path that it saves to}
}
\value{
Returns NULL
}
\description{
A helper function that saves alignment to a fasta file
}
\examples{
# Example 1
# Create a basic msa and then plot a tree from it
package(msa)
package(Biostrings)
set1 <- Biostrings::DNAStringSet("ATCGATCG")
set2 <- Biostrings::DNAStringSet("ATTTTTTT")
set3 <- Biostrings::DNAStringSet("ATCGATTT")
set <- union(set1, set2)
set <- union(set, set3)
align <- multiplSeqAlign(set)
saveAlignmentToFasta(align, "align.fasta")
\dontrun{
# Example 2
}
}
\references{
Charif D, Lobry J. 2007. “SeqinR 1.0-2: a contributed package to the R
project for statistical computing devoted to biological sequences retrieval
and analysis.” In Bastolla U, Porto M, Roman H, Vendruscolo M (eds.),
Structural approaches to sequence evolution: Molecules, networks,
populations, series Biological and Medical Physics, Biomedical Engineering,
207-232. Springer Verlag, New York.
U. Bodenhofer, E. Bonatesta, C. Horejs-Kainrath, and S. Hochreiter (2015) msa:
an R package for multiple sequence alignment. Bioinformatics 31(24):3997-
9999. DOI: 10.1093/bioinformatics/btv176.
}
