% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simpleAlign}
\alias{simpleAlign}
\title{Result from alignment of few simple sequences}
\format{
A MsaAAMultipleAlignment Object:
}
\usage{
simpleAlign
}
\description{
A dataset contains alignmentg of three short sequences
}
\examples{
\dontrun{
 simpleAlign
}
}
\keyword{datasets}
