% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{unionDNASets}
\alias{unionDNASets}
\title{Union two DNAStringSet}
\usage{
unionDNASets(set1, set2)
}
\arguments{
\item{set1}{A DNAStringSet contains DNA sequence}

\item{set2}{A DNAStringSet contains DNA sequence}
}
\value{
Returns a DNAStringSet with sequences contained in set1 and set2.
}
\description{
A function that takes two DNAStringSet and union them into one.
}
\examples{
# Example 1
library(Biostrings)
set1 <- Biostrings::DNAStringSet("ATCGATCG")
set2 <- Biostrings::DNAStringSet("ATTTTTTT")
set3 <- Biostrings::DNAStringSet("ATCGATTT")
setTotal <- unionDNASets(set1, set2)

}
\references{
H. Pagès, P. Aboyoun, R. Gentleman and S. DebRoy (2020). Biostrings:
Efficient manipulation of biological strings.
R package version 2.58.0. https://bioconductor.org/packages/Biostrings

Paradis E. & Schliep K. 2019. ape 5.0: an environment for modern
phylogenetics and evolutionary
analyses in R. Bioinformatics 35: 526-528.
}
