% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeneticTree.R
\name{createTree}
\alias{createTree}
\title{Create a neighbor join tree object}
\usage{
createTree(alignment)
}
\arguments{
\item{alignment}{A MsaAAMultipleAlignment object for creating phylogenetic
tree}
}
\value{
Returns an phylo object which is a phylogenetic tree
}
\description{
A function that creates a neighbor join tree from multiple sequence alignment
}
\examples{
# Example 1
# Using GeneCounts dataset available with package
dim(GeneCounts)

# Calculate information criteria value
InfCriteriaResults <- InfCriteriaCalculation(loglikelihood = -5080,
                                             nClusters = 2,
                                             dimensionality = ncol(GeneCounts),
                                             observations = nrow(GeneCounts),
                                             probability = c(0.5, 0.5))
# To obtain BIC value from results
InfCriteriaResults$BICresults

}
\references{
Charif D, Lobry J. 2007. “SeqinR 1.0-2: a contributed package to the R
project for statistical computing devoted to biological sequences retrieval
and analysis.” In Bastolla U, Porto M, Roman H, Vendruscolo M (eds.),
Structural approaches to sequence evolution: Molecules, networks,
populations, series Biological and Medical Physics, Biomedical Engineering,
207-232. Springer Verlag, New York.

Paradis E. & Schliep K. 2019. ape 5.0: an environment for modern
phylogenetics and evolutionaryanalyses in R. Bioinformatics 35: 526-528.

U. Bodenhofer, E. Bonatesta, C. Horejs-Kainrath, and S. Hochreiter (2015) msa:
an R package for multiple sequence alignment. Bioinformatics 31(24):3997-
9999. DOI: 10.1093/bioinformatics/btv176.
}
