% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceComparison.R
\name{plotAlignment}
\alias{plotAlignment}
\title{Compare multiple sequence alignments}
\usage{
plotAlignment(alignment, outputName = "align.fasta")
}
\arguments{
\item{sequences}{A AAStringset containing several sequences}

\item{algorithm}{A string indicating the algorithm that user wants to use
to calculate multiple sequence alignment}
}
\value{
Returns a msa
}
\description{
A function that calculates multiple sequence alignments
}
\examples{
# Example 1
# Create a basic msa and then plot a tree from it
package(Biostrings)
set1 <- Biostrings::AAStringSet("ATCGATCG")
set2 <- Biostrings::AAStringSet("ATTTTTTT")
set3 <- Biostrings::AAStringSet("ATCGATTT")
set <- union(set1, set2)
set <- union(set, set3)
align <- multiplSeqAlign(set)
plotAlignment(align)

\dontrun{
# Example 2
}
}
\references{
Charif D, Lobry J. 2007. “SeqinR 1.0-2: a contributed package to the R
project for statistical computing devoted to biological sequences retrieval
and analysis.” In Bastolla U, Porto M, Roman H, Vendruscolo M (eds.),
Structural approaches to sequence evolution: Molecules, networks,
populations, series Biological and Medical Physics, Biomedical Engineering,
207-232. Springer Verlag, New York.

Venket Raghavan (2021). seqvisr: Biological Sequence Visualization Functions
in R. R package version 0.2.5. https://github.com/vragh/seqvisr

U. Bodenhofer, E. Bonatesta, C. Horejs-Kainrath, and S. Hochreiter (2015) msa:
an R package for multiple sequence alignment. Bioinformatics 31(24):3997-
9999. DOI: 10.1093/bioinformatics/btv176.
}
